--- tests/Refinery/Null/GroupTest.php
@@ -1,48 +1,0 @@
-<?php
-/* Copyright (c) 1998-2019 ILIAS open source, Extended GPL, see docs/LICENSE */
-
-namespace ILIAS\Tests\Refinery\Null;
-
-use ILIAS\Data\Factory;
-use ILIAS\Refinery\Integer\GreaterThan;
-use ILIAS\Refinery\Integer\LessThan;
-use ILIAS\Refinery\Null\IsNull;
-use ILIAS\Refinery\Null\Group;
-use ILIAS\Tests\Refinery\TestCase;
-
-require_once('./libs/composer/vendor/autoload.php');
-
-class GroupTest extends TestCase
-{
-	/**
-	 * @var Group
-	 */
-	private $group;
-
-	/**
-	 * @var Factory
-	 */
-	private $dataFactory;
-
-	/**
-	 * @var \ilLanguage
-	 */
-	private $language;
-
-
-	public function setUp() : void
-	{
-		$this->dataFactory = new Factory();
-		$this->language    = $this->getMockBuilder('\ilLanguage')
-			->disableOriginalConstructor()
-			->getMock();
-
-		$this->group = new Group($this->dataFactory, $this->language);
-	}
-
-	public function testIsNullGroup()
-	{
-		$instance = $this->group->isNull();
-		$this->assertInstanceOf(IsNull::class, $instance);
-	}
-}
